rename 'pelo' -> 'por'
