+++ src/repositorios/volatil/armazenamento_usuario_volatil.py
@@ -143,7 +143,7 @@ class ArmazenamentoUsuarioVolatil(IArmazenamento, IAlteracaoInfosCadastro, IDele
             return u[0]
         return []
     
-    def deletarUsuarioPeloEmail(self, email: str):
+    def deletarUsuarioPorEmail(self, email: str):
         for u in self.armazem:
             for e in u.contato.emails:
                 if e.email == email:
