+++ src/usecases/uc_deletar_usuario_por_email.py
@@ -14,4 +14,4 @@ class UCDeletarUsuarioPorEmail():
         if not(self.deletarUsuario.usuarioExistePorEmail(email)):
             raise ErroUsuarioInvalido
         
-        self.deletarUsuario.deletarUsuarioPeloEmail(email)
+        self.deletarUsuario.deletarUsuarioPorEmail(email)
