+++ tests/controladores/test_controller_deletar_usuario_por_email_fastapi.py
@@ -1,4 +1,4 @@
-from src.controladores.control_deletar_usuario_pelo_email_fastapi import ControllerHTTPDeletarUsuarioPorEmailFastAPI
+from src.controladores.control_deletar_usuario_por_email_fastapi import ControllerHTTPDeletarUsuarioPorEmailFastAPI
 from src.repositorios.volatil.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
 from src.usecases.uc_deletar_usuario_por_email import UCDeletarUsuarioPorEmail
 from src.usecases.uc_cadastrar_usuario import UCCadastrarUsuario
